Add release workflow
Since the latest version of RubyGems supports passing OTP codes via an
environment variable, this means releases are a bit easier and can be
automated via a GitHub Actions workflow.

Ideally I would extend this to autogenerate the CHANGELOG file and
create a GitHub Release based on the tag as well.